+++ examples/bird-checker-with-nextjs-and-eval/CHANGELOG.md
@@ -1,5 +1,19 @@
 # bird-checker-with-nextjs-and-eval
 
+## 0.1.1-alpha.33
+
+### Patch Changes
+
+- Updated dependencies [d2cd535]
+  - @mastra/core@0.1.27-alpha.54
+
+## 0.1.1-alpha.32
+
+### Patch Changes
+
+- Updated dependencies [8e7814f]
+  - @mastra/core@0.1.27-alpha.53
+
 ## 0.1.1-alpha.31
 
 ### Patch Changes