+++ examples/crypto-chatbot/CHANGELOG.md
@@ -1,5 +1,21 @@
 # ai-chatbot
 
+## 0.1.1-alpha.34
+
+### Patch Changes
+
+- Updated dependencies [d2cd535]
+  - @mastra/core@0.1.27-alpha.54
+  - @mastra/memory@0.0.2-alpha.34
+
+## 0.1.1-alpha.33
+
+### Patch Changes
+
+- Updated dependencies [8e7814f]
+  - @mastra/core@0.1.27-alpha.53
+  - @mastra/memory@0.0.2-alpha.33
+
 ## 0.1.1-alpha.32
 
 ### Patch Changes