+++ examples/dane/CHANGELOG.md
@@ -1,5 +1,33 @@
 # @mastra/dane
 
+## 0.0.2-alpha.46
+
+### Patch Changes
+
+- Updated dependencies [eb45d76]
+- Updated dependencies [d2cd535]
+  - @mastra/rag@0.0.2-alpha.38
+  - @mastra/core@0.1.27-alpha.54
+  - @mastra/firecrawl@1.0.4-alpha.40
+  - @mastra/github@1.0.3-alpha.38
+  - @mastra/stabilityai@1.0.1-alpha.29
+  - @mastra/engine@0.0.5-alpha.48
+  - @mastra/memory@0.0.2-alpha.34
+
+## 0.0.2-alpha.45
+
+### Patch Changes
+
+- 8e7814f: Add payload getter on machine context
+- Updated dependencies [8e7814f]
+  - @mastra/core@0.1.27-alpha.53
+  - @mastra/firecrawl@1.0.4-alpha.39
+  - @mastra/github@1.0.3-alpha.37
+  - @mastra/stabilityai@1.0.1-alpha.28
+  - @mastra/engine@0.0.5-alpha.47
+  - @mastra/memory@0.0.2-alpha.33
+  - @mastra/rag@0.0.2-alpha.37
+
 ## 0.0.2-alpha.44
 
 ### Patch Changes