+++ integrations/ashby/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @mastra/ashby
 
+## 1.0.1-alpha.33
+
+### Patch Changes
+
+- Updated dependencies [d2cd535]
+  - @mastra/core@0.1.27-alpha.54
+
+## 1.0.1-alpha.32
+
+### Patch Changes
+
+- Updated dependencies [8e7814f]
+  - @mastra/core@0.1.27-alpha.53
+
 ## 1.0.1-alpha.31
 
 ### Patch Changes