+++ integrations/firecrawl/CHANGELOG.md
@@ -1,5 +1,22 @@
 # @mastra/firecrawl
 
+## 1.0.4-alpha.40
+
+### Patch Changes
+
+- Updated dependencies [eb45d76]
+- Updated dependencies [d2cd535]
+  - @mastra/rag@0.0.2-alpha.38
+  - @mastra/core@0.1.27-alpha.54
+
+## 1.0.4-alpha.39
+
+### Patch Changes
+
+- Updated dependencies [8e7814f]
+  - @mastra/core@0.1.27-alpha.53
+  - @mastra/rag@0.0.2-alpha.37
+
 ## 1.0.4-alpha.38
 
 ### Patch Changes