+++ packages/core/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @mastra/core
 
+## 0.1.27-alpha.54
+
+### Patch Changes
+
+- d2cd535: configure dotenv in core
+
+## 0.1.27-alpha.53
+
+### Patch Changes
+
+- 8e7814f: Add payload getter on machine context
+
 ## 0.1.27-alpha.52
 
 ### Patch Changes