+++ packages/engine/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @mastra/engine
 
+## 0.0.5-alpha.48
+
+### Patch Changes
+
+- Updated dependencies [d2cd535]
+  - @mastra/core@0.1.27-alpha.54
+
+## 0.0.5-alpha.47
+
+### Patch Changes
+
+- Updated dependencies [8e7814f]
+  - @mastra/core@0.1.27-alpha.53
+
 ## 0.0.5-alpha.46
 
 ### Patch Changes