+++ packages/rag/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @mastra/rag
 
+## 0.0.2-alpha.38
+
+### Patch Changes
+
+- eb45d76: Updated PG Vector filter and added rag examples in docs
+- Updated dependencies [d2cd535]
+  - @mastra/core@0.1.27-alpha.54
+
+## 0.0.2-alpha.37
+
+### Patch Changes
+
+- Updated dependencies [8e7814f]
+  - @mastra/core@0.1.27-alpha.53
+
 ## 0.0.2-alpha.36
 
 ### Patch Changes